Change of the structure of TableWidok
new class DruzynyWidok
changable panels
there is no new window for matches
