+++ TabelaWidok.java
@@ -1,7 +1,4 @@
 import javax.swing.*;
-import javax.swing.plaf.basic.BasicScrollBarUI;
-import javax.swing.table.DefaultTableCellRenderer;
-import javax.swing.table.JTableHeader;
 import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
@@ -10,6 +7,10 @@ import java.net.URL;
 import javax.imageio.ImageIO;
 
 class TabelaWidok extends JFrame {
+
+    CardLayout uklad_kart = new CardLayout();
+    JPanel panel = new JPanel(uklad_kart);
+
     TabelaWidok(Tabela tabela) {
         UIManager.put("TextField.selectionBackground", new Color(124, 252, 0));
         UIManager.put("TextArea.selectionBackground", new Color(124, 252, 0));
@@ -24,93 +25,13 @@ class TabelaWidok extends JFrame {
         setTitle("Tabela drużyn");
         getContentPane().setBackground(new Color(0, 100, 0));
 
-        String[] nazwy_kolumn = {"Klub", "Pkt", "RM", "W", "R", "P", "BZ", "BS"};
+        JPanel panel_druzyn = new DruzynyWidok(tabela);
+        JPanel panel_meczy = new MeczeWidok(tabela);
 
-        Object[][] dane = new Object[tabela.druzyny.size()][8];
-        for (int i = 0; i < tabela.druzyny.size(); i++) {
-            Druzyna d = tabela.druzyny.get(i);
-            dane[i][0] = d.getNazwa();
-            dane[i][1] = d.getPunkty();
-            dane[i][2] = d.getRozegraneMecze();
-            dane[i][3] = d.getWygrane();
-            dane[i][4] = d.getRemisy();
-            dane[i][5] = d.getPrzegrane();
-            dane[i][6] = d.getBramkiZdobyte();
-            dane[i][7] = d.getBramkiStracone();
-        }
+        panel.add(panel_druzyn, "Drużyny");
+        panel.add(panel_meczy, "Mecze");
 
-        JTable wiersze = new JTable(dane, nazwy_kolumn);
-        wiersze.setRowHeight(30);
-        wiersze.setShowVerticalLines(false);
-        ((DefaultTableCellRenderer) wiersze.getDefaultRenderer(Object.class)).setBorder(BorderFactory.createLineBorder(Color.WHITE));
-
-        DefaultTableCellRenderer centerRenderer = new DefaultTableCellRenderer();
-        centerRenderer.setHorizontalAlignment(JLabel.CENTER);
-        for (int i = 1; i < wiersze.getColumnCount(); i++) {
-            wiersze.getColumnModel().getColumn(i).setCellRenderer(centerRenderer);
-        }
-
-        wiersze.setBackground(new Color(0, 100, 0));
-        wiersze.setForeground(Color.WHITE);
-
-        JTableHeader header = wiersze.getTableHeader();
-        header.setFont(header.getFont().deriveFont(Font.BOLD));
-
-        ((DefaultTableCellRenderer) header.getDefaultRenderer()).setBorder(BorderFactory.createEmptyBorder());
-
-        header.setBackground(new Color(0, 100, 0));
-        header.setForeground(Color.WHITE);
-
-        JPanel panel = new JPanel(new BorderLayout());
-        panel.setBackground(new Color(0, 100, 0));
-
-        JScrollBar przewijak = new JScrollBar() {
-            @Override
-            public void updateUI() {
-                setUI(new BasicScrollBarUI() {
-                    @Override
-                    protected void paintThumb(Graphics g, JComponent c, Rectangle thumbBounds) {
-                        g.setColor(Color.WHITE);
-                        g.fillRect(thumbBounds.x, thumbBounds.y, thumbBounds.width, thumbBounds.height);
-                    }
-
-                    @Override
-                    protected void paintTrack(Graphics g, JComponent c, Rectangle trackBounds) {
-                        g.setColor(new Color(0, 100, 0));
-                        g.fillRect(trackBounds.x, trackBounds.y, trackBounds.width, trackBounds.height);
-                    }
-
-                    @Override
-                    protected JButton createDecreaseButton(int orientation) {
-                        return brak_przycisku();
-                    }
-
-                    @Override
-                    protected JButton createIncreaseButton(int orientation) {
-                        return brak_przycisku();
-                    }
-
-                    private JButton brak_przycisku() {
-                        JButton button = new JButton();
-                        button.setPreferredSize(new Dimension(0, 0));
-                        button.setMinimumSize(new Dimension(0, 0));
-                        button.setMaximumSize(new Dimension(0, 0));
-                        return button;
-                    }
-                });
-            }
-        };
-
-        JPanel naroznik = new JPanel();
-        naroznik.setBackground(new Color(0, 100, 0));
-
-        JScrollPane przewijalne = new JScrollPane(wiersze);
-        przewijalne.setCorner(JScrollPane.UPPER_RIGHT_CORNER, naroznik);
-        przewijalne.setVerticalScrollBar(przewijak);
-        przewijalne.getViewport().setBackground(new Color(0, 100, 0));
-        panel.add(przewijalne, BorderLayout.CENTER);
-        panel.setSize(300, 150);
-        panel.setVisible(true);
+        uklad_kart.show(panel, "Drużyny");
 
         JPanel gorny_panel = new JPanel();
         gorny_panel.setLayout(new BorderLayout());
@@ -137,11 +58,25 @@ class TabelaWidok extends JFrame {
         druzyny.setFont(new Font("Dialog", Font.BOLD, 16));
         gorne_przyciski.add(druzyny);
 
+        druzyny.addActionListener(new ActionListener() {
+            @Override
+            public void actionPerformed(ActionEvent e) {
+                uklad_kart.show(panel, "Drużyny");
+            }
+        });
+
         JButton mecze = new JButton("Mecze");
         mecze.setBackground(new Color(0, 100, 0));
         mecze.setForeground(Color.WHITE);
         mecze.setFont(new Font("Dialog", Font.BOLD, 16));
         gorne_przyciski.add(mecze);
+
+        mecze.addActionListener(new ActionListener() {
+            @Override
+            public void actionPerformed(ActionEvent e) {
+                uklad_kart.show(panel, "Mecze");
+            }
+        });
 
         gorny_panel.add(gorne_przyciski, BorderLayout.SOUTH);
 
@@ -156,17 +91,9 @@ class TabelaWidok extends JFrame {
         dolny_panel.setBackground(new Color(0, 100, 0));
         dolny_panel.setForeground(Color.WHITE);
 
-        panel.add(gorny_panel, BorderLayout.NORTH);
-        panel.add(dolny_panel, BorderLayout.SOUTH);
-
-        add(panel);
-
-        mecze.addActionListener(new ActionListener() {
-            @Override
-            public void actionPerformed(ActionEvent e) {
-                new MeczeWidok(tabela);
-            }
-        });
+        add(gorny_panel, BorderLayout.NORTH);
+        add(panel, BorderLayout.CENTER);
+        add(dolny_panel, BorderLayout.SOUTH);
 
         setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
